remove unused code and refactor tests in client
