+++ client/src/tests/load_chain_tests.rs
@@ -5,13 +5,18 @@
 extern crate tempdir;
 
 use self::tempdir::TempDir;
-use crate::archive::{ArchiveClient, Configuration};
+use crate::{
+    archive::{ArchiveClient, Configuration},
+    rpc::RpcBlock,
+};
 use cfx_types::H256;
 use parking_lot::{Condvar, Mutex};
+use primitives::Block;
 use serde_json::Value;
 use std::{
     fs::File,
-    io::Read,
+    io::{BufReader, Read},
+    path::Path,
     sync::Arc,
     thread,
     time::{Duration, Instant},
@@ -65,13 +70,37 @@ fn test_load_chain() {
             .into_string()
             .unwrap(),
     );
-    conf.raw_conf.load_test_chain =
-        Some(r#"../tests/blockchain_tests/general_2.json"#.to_owned());
     conf.raw_conf.port = Some(13000);
     conf.raw_conf.jsonrpc_http_port = Some(18000);
 
     let exit = Arc::new((Mutex::new(false), Condvar::new()));
     let handle = ArchiveClient::start(conf, exit.clone()).unwrap();
+
+    let chain_path = "../tests/blockchain_tests/general_2.json";
+
+    // make sure db recovery has completed
+    thread::sleep(Duration::from_secs(7));
+
+    let file_path = Path::new(&chain_path);
+    let file = File::open(file_path)
+        .map_err(|e| format!("Failed to open test-chain file {:?}", e))
+        .ok()
+        .unwrap();
+    let reader = BufReader::new(file);
+    let rpc_blocks: Vec<RpcBlock> = serde_json::from_reader(reader)
+        .map_err(|e| format!("Failed to parse blocks from json {:?}", e))
+        .ok()
+        .unwrap();
+    assert!(
+        !rpc_blocks.is_empty(),
+        "Error: The json data should not be empty."
+    );
+    for rpc_block in rpc_blocks.into_iter().skip(1) {
+        let primitive_block: Block = rpc_block.into_primitive().map_err(|e| {
+            format!("Failed to convert from a rpc_block to primitive block {:?}", e)
+        }).ok().unwrap();
+        handle.sync.on_mined_block(primitive_block);
+    }
 
     let expected = get_expected_best_hash();
     let best_block_hash: H256 =
