+++ tests/crash_archive_era150_test.py
@@ -18,7 +18,6 @@ class CrashArchiveNodeTest(ConfluxTestFramework):
     def set_test_params(self):
         self.setup_clean_chain = True
         self.num_nodes = 8
-        self.conf_parameters = {"persist_terminal_period_ms": "500"}
         self.conf_parameters["log_level"] = "\"debug\""
         self.conf_parameters["era_epoch_count"] = "150"
         self.conf_parameters["era_checkpoint_gap"] = "150"
