+++ tests/crash_archive_test.py
@@ -18,7 +18,6 @@ class CrashFullNodeTest(ConfluxTestFramework):
     def set_test_params(self):
         self.setup_clean_chain = True
         self.num_nodes = 8
-        self.conf_parameters = {"persist_terminal_period_ms": "500"}
 
     def setup_network(self):
         self.add_nodes(self.num_nodes)
